\name{data}
\alias{data}
\docType{data}
\title{
Example data for MAVTn
}
\description{
a data matrix with 15 samples in column and one class type and 10 gene expression data in row. 
}
\usage{data(data)}
\format{
  The format is:
 num [1:16, 1:10] 1 -0.0853 1.0868 -0.4995 0.9477 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:16] "cl" "" "" "" ...
  ..$ : NULL
}

