\name{design.matrix}
\alias{design.matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Design matrix
}
\description{
To construct a design matrix of the clinical outcome of the samples.
}
\usage{
design.matrix(factors)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{factors}{
The clinical outcome of the samples.
}
}

\value{
A design matrix is returned.
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


