\name{ma.estimate}
\alias{ma.estimate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimate of the coefficients
}
\description{
To calculate the ordinary least square estimate of the coefficients. 
}
\usage{
ma.estimate(Y, X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Y}{
Data matrix; row is sample; each column is variable(gene). 
}
  \item{X}{
Design matrix.
}
}

\value{
The estimate of the coefficients is returned.
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.


