\name{MAVTgsa-package}
\alias{MAVTgsa-package}
\alias{MAVTgsa}
\docType{package}
\title{
OLS and Multivariate Analysis of Variance test for gene set analysis
}
\description{
A gene set analysis function for one-sided test (OLS) and two-sided test (multivariate analysis of variance).
Three multiple comparison procedures, Dunnett, Tukey, and sequential pairwise comparison, are implemented. 
MAVTgsa computes the p-values and FDR (false discovery rate) q-values for all gene sets.
The p-values for individual genes in a significant gene set are also listed.
MAVTgsa generates a GST-plot of the empirical distribution function of the ranked test statistics of a given gene set.
}
\details{
\tabular{ll}{
Package: \tab MAVTgsa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-04\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Chih-Yi Chien, Chen-An Tsai, Ching-Wei Chang, and James J. Chen

Maintainer: Chih-Yi Chien <chihyi.chien@fda.hhs.gov>
}
\references{
Chen,J.J. et al. (2007) Significance analysis of group of genes in expression profiling studies, Bioinformatics, 23, 2104.

Tsai,C.A. et al. (2009) Multivariate analysis of variance test for gene set analysis. Bioinformatics, 25, 897.
}

\keyword{ one-sided test, MANOVA test, multiple comparisons, OLS test, gene set analysis, GST-plot }
