\name{MAVTgsa-package}
\alias{MAVTgsa-package}
\alias{MAVTgsa}
\docType{package}
\title{
OLS and Multivariate Analysis of Variance test for gene set analysis
}
\description{
             A gene set analysis functions for one-sided test (OLS) and two-sided test (multivariate analysis of variance).
             If the experimental conditions are equal to 2, the p-value for Hotelling's t^2 test is calculated.
             If the experimental conditions are great than 2, the p-value for Wilks' Lambda is deterimined and post-hoc test is reported too.
             The p-value for individual gene test of significant gene sets are also listed.}
\details{
\tabular{ll}{
Package: \tab MAVTgsa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-01\cr
License: \tab No commercial use\cr
}

}
\author{
Chih-Yi Chien, Chen-An Tsai, Ching-Wei Chang, and James J. Chen

Maintainer: Chih-Yi Chien <chihyi.chien@fda.hhs.gov>
}
\references{
Chen,J.J. et al. (2007) Significance analysis of group of genes in expression profiling studies, Bioinformatics, 23, 2104.

Tsai,C.A. et al. (2009) Multivariate analysis of variance test for gene set analysis. Bioinformatics, 25, 897.
}

\keyword{ gene set analysis, OLS, Hotelling's T2, MANOVA }

